+++ TMessagesProj/src/main/java/org/telegram/ui/Components/TranslateButton.java
@@ -56,7 +56,7 @@ public class TranslateButton extends FrameLayout {
     private final boolean[] accusative = new boolean[1];
 
     public TranslateButton(Context context, ChatActivity chatActivity, Theme.ResourcesProvider resourcesProvider) {
-        this(context, chatActivity.getCurrentAccount(), chatActivity.getDialogId(), chatActivity.getTopicId(), chatActivity, resourcesProvider);
+        this(context, chatActivity.getCurrentAccount(), chatActivity.getDialogId(), (int) chatActivity.getTopicId(), chatActivity, resourcesProvider);
     }
 
     public TranslateButton(Context context, int currentAccount, long dialogId, int topicId, BaseFragment fragment, Theme.ResourcesProvider resourcesProvider) {
